Update to Scipy v1.1 (#3568)
* Fix (most) MeasureImageOverlap errors

* Fix Align for scipy v1.1

* Fix other MeasureImageOverlap tests by adjusting targets

* Update to v1.1 of scipy

* FIXME: Remove failing test due to upstream scipy issues

* Revert addition of uint32 dtype in coo matrix (#3590)

* Rebased3565 (#3591)

* use bilateral filter from skimage (#3567)

* use bilateral filter from skimage

* use bilateral filter from skimage

* use bilateral filter from skimage

* Implictly rescale volumes (#3570)

* Implictly rescale volumes

* Rescale numpy vol

* fix test name

* Make ball uint16 not float64

* Rescale expected test

* Rescale expected test image

* Test almost equal, not equal

* Remove references to 'highlevel'

* Update hdf5_dict.py

* skimage 0.14 fixes (#3580)

* Update setup.py

* Change from marching_cubes to marching_cubes_classic

* marching_cubes to marching_cubes_classic

* marching_cubes_classic only returns verts and faces

* marching_cubes_classic only returns verts and faces

* Use scipy binary_erosion (as elsewhere)

* skimage 0.14 fixes (#3580) (#3584)

* Update setup.py

* Change from marching_cubes to marching_cubes_classic

* marching_cubes to marching_cubes_classic

* marching_cubes_classic only returns verts and faces

* marching_cubes_classic only returns verts and faces

* Use scipy binary_erosion (as elsewhere)

* h5py now really no longer 2.8 (#3585)

* Issues/3531 (#3577)

* adds the test to reproduce issue #3531

* image intensity rescaled to -1 to 1 while resizing

* improved the test

* small edit in moduleview.py fixes #3568 (#3587)

* Documentation overhaul (#3563)

* Add standard text for structuring elements

* Expanded Advanced module documentation

* Resolves #3503

* Resolves #3504

* Resolves #3430

* Resolves #3326

* Reolves #3552 and #3559

* Typos and year fix

* Update matchtemplate.py

* Update medialaxis.py

* Update watershed.py

* Remove double spaces

* Explain 2D long

* Fix MedianFilter docs

* Add structuring element diagram.

* Style fixes

* Update matchtemplate.py

+++ cellprofiler/modules/measureimageoverlap.py
@@ -511,8 +511,8 @@ the two images. Set this setting to “No” to assess no penalty."""
 
         returns a tuple of the Rand Index and the adjusted Rand Index
         """
-        ground_truth_labels = ground_truth_labels[mask].astype(numpy.uint64)
-        test_labels = test_labels[mask].astype(numpy.uint64)
+        ground_truth_labels = ground_truth_labels[mask].astype(numpy.uint32)
+        test_labels = test_labels[mask].astype(numpy.uint32)
         if len(test_labels) > 0:
             #
             # Create a sparse matrix of the pixel labels in each of the sets
